checkMail UI
