+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mboacare/app_modules/auth/forgotPassword.dart';
-import 'package:mboacare/app_modules/auth/resetPassword.dart';
+import 'package:mboacare/app_modules/auth/checkMail.dart';
 import 'package:mboacare/app_modules/med_user/screen/dashboard/home.dart';
 import 'package:mboacare/app_modules/med_user/screen/dashboard/hospital/awaitApprovalScreen.dart';
 import 'package:mboacare/app_modules/med_user/screen/inner_screen/user_profile_page.dart';
@@ -108,7 +108,7 @@ class MyApp extends StatelessWidget {
         ('/themeScreen'): (context) => ThemeScreen(),
         ('/profilePage'): (context) => ProfilePage(),
         ('/awaitApproval'): (context) => AwaitApprovalScreen(),
-        ('/resetPassword'): (context) => ResetPasswordScreen(),
+        ('/resetPassword'): (context) => CheckMailScreen(),
         ('/forgotPassword'): (context) => ForgotPasswordScreen()
       },
     );
